// src/app/models/orderModel.ts
export interface OrderModel {
  payment_method: string;
  payment_method_title: string;
  set_paid: boolean;
  billing: {
    first_name: string;
    last_name: string;
    address_1: string;
    address_2: string;
    city: string;
    state: string;
    postcode: string;
    country: string;
    email: string;
    phone: string;
  };
  shipping: {
    first_name: string;
    last_name: string;
    address_1: string;
    address_2: string;
    city: string;
    state: string;
    postcode: string;
    country: string;
  };
  line_items: LineItemsModel[];
  customer_id?: number;
}

export interface LineItemsModel {
  product_id: number;
  quantity: number;
  variation_id?: number;
}